T38 home page bugs fixed
